requirements.txt
rich
tqdm
wandb
pytorch-lightning
torchmetrics
transformers
sklearn
datasets
sentencepiece
tensorboard
fairscale
deepspeed
bitsandbytes
accelerate
